+++ src/main/java/com/kau/capstone/domain/pet/util/PetMapper.java
@@ -3,7 +3,7 @@ package com.kau.capstone.domain.pet.util;
 import com.kau.capstone.domain.pet.dto.request.PetRegistRequest;
 import com.kau.capstone.domain.pet.dto.response.PetInfoResponse;
 import com.kau.capstone.domain.pet.entity.Pet;
-import com.kau.capstone.global.common.Gender;
+import com.kau.capstone.domain.pet.entity.Gender;
 import com.kau.capstone.domain.pet.entity.PetType;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
